+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#262](https://github.com/babylonlabs-io/finality-provider/pull/262) Add new command to export pop
 * [#284](https://github.com/babylonlabs-io/finality-provider/pull/284) Add new command to delete pop
 * [#277](https://github.com/babylonlabs-io/finality-provider/pull/277) Poll many blocks in poller
+* [#291](https://github.com/babylonlabs-io/finality-provider/pull/291) chore: remove skip height
 
 ## v0.14.3